created a Home componet, set up react-router-dom in src/App.js to render Home component, also cleaned up some code
