+++ shopping-cart-app/src/App.js
@@ -1,46 +1,11 @@
-import React, { Component } from "react";
-import StoreCard from "./components/StoreCard";
-import Wrapper from "./components/Wrapper";
-import Header from "./components/Header";
-import friends from "./friends.json";
-import "./App.css";
+import React from 'react';
+import { BrowserRouter as Router, Route } from 'react-router-dom';
+import Home from './components/Home';
 
-class App extends Component {
-  // Setting this.state.friends to the friends json array
-  state = {
-    friends
-  };
-
-  removeStore = id => {
-    // Filter this.state.friends for friends with an id not equal to the id being removed
-    const friends = this.state.friends.filter(friend => friend.id !== id);
-    // Set this.state.friends equal to the new friends array
-    this.setState({ friends });
-  };
-
-  // Map over this.state.friends and render a FriendCard component for each friend object
-  render() {
-    return (
-     <div>
-        <Header location="Search all stores"/>
-         <Wrapper>
-        {this.state.friends.map(friend => 
-          <StoreCard
-            removeStore={this.removeStore}
-            id={friend.id}
-            key={friend.id}
-            name={friend.name}
-            storeImage={friend.storeImage}
-            image1={friend.image1}
-            image2={friend.image2}
-            image3={friend.image3}
-            description={friend.description}
-          />
-        )}
-      </Wrapper>
-      </div>
-    );
-  }
-}
+const App = () => (
+  <Router>
+    <Route path="/" component={Home} />
+  </Router>
+);
 
 export default App;
