+++ shopping-cart-app/src/components/StoreCard/index.js
@@ -1,1 +1,1 @@
-export { default } from "./StoreCard";
+export { default } from './StoreCard';
