+++ shopping-cart-app/src/components/Wrapper/Wrapper.js
@@ -1,5 +1,5 @@
-import React from "react";
-import "./Wrapper.css";
+import React from 'react';
+import './Wrapper.css';
 
 const Wrapper = props => <div className="wrapper">{props.children}</div>;
