+++ shopping-cart-app/src/components/Wrapper/index.js
@@ -1,1 +1,1 @@
-export { default } from "./Wrapper";
+export { default } from './Wrapper';
