+++ start-shopping-cart-app.js
@@ -1,3 +1,3 @@
-const args = [ 'start' ];
+const args = ['start'];
 const opts = { stdio: 'inherit', cwd: 'shopping-cart-app', shell: true };
 require('child_process').spawn('npm', args, opts);
